每日三次投票逻辑修正

+++ cloudfunctions/delArticle/index.js
@@ -1,7 +1,7 @@
 // 云函数入口文件
 const cloud = require('wx-server-sdk')
 
-cloud.init({env: cloud.DYNAMIC_CURRENT_ENV})
+cloud.init({env: "test-8gz67lpof2b9185f"})
 const db = cloud.database()
 
 // 云函数入口函数
@@ -12,8 +12,16 @@ exports.main = async (event, context) => {
       await transaction.collection("vote").where(event).remove()
       // 获取作品实例，从中获得图片ID字段
       let res = await transaction.collection("article").doc(event.articleId).get()
+      if (!res.data) {
+        throw new Error("查询作品错误！返回值缺少data字段")
+      }
       // 删除图片文件
       await cloud.deleteFile({fileList: [res.data.picID]})
+      // 修改选手状态为未参赛
+      await transaction.collection("player").where({
+        name: res.data.author,
+        room: res.data.room
+      }).update({data: {status: "未参赛"}})
       // 删除作品
       await transaction.collection("article").doc(event.articleId).remove()
     } catch(e) {
